requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.6.1
PyJWT==2.11.0
Werkzeug==3.1.6
eventlet==0.40.3
gunicorn==23.0.0
mysql-connector-python==9.6.0
mysqlclient==2.2.8
qrcode==8.2
